requirements.txt
ansi2html==1.9.2
PyQt6==6.7.0
PyQt6-Qt6==6.7.2
PyQt6-sip==13.6.0
pyserial==3.5
